integrate bibliography into bims core
