+++ bims/models/__init__.py
@@ -11,3 +11,4 @@ from bims.models.boundary import *  # noqa
 from bims.models.boundary_type import *  # noqa
 from bims.models.carousel_header import CarouselHeader
 from bims.models.links import *
+from bims.models.bibliography import * #noqa
