+++ bims/templates/td_biblio/entry_item.html
