+++ bims/templates/td_biblio/entry_list.html
