+++ bims/templates/td_biblio/partials/publication_list_by_year.html
@@ -12,7 +12,7 @@
         <ul class="publication-list">
         {% for entry in entry_year.list %}
             <li class="publication">
-                {% include "td_biblio/entry_item.html" %}
+                {% include "td_biblio/../td_biblio/entry_item.html" %}
             </li>
         {% endfor %}
         </ul>
