+++ core/settings/contrib.py
@@ -89,9 +89,11 @@ CELERY_ACCEPT_CONTENT = ['json']
 CELERY_TASK_SERIALIZER = 'json'
 CELERY_RESULT_SERIALIZER = 'json'
 
-BROKER_URL = 'amqp://guest:guest@%s:5672//' % os.environ['RABBITMQ_HOST']
+# BROKER_URL = 'amqp://guest:guest@%s:5672//' % os.environ['RABBITMQ_HOST']
+BROKER_URL = 'amqp://guest:guest@rabbitmq:5672//'
 
 # contact us email
-CONTACT_US_EMAIL = os.environ['CONTACT_US_EMAIL']
+# CONTACT_US_EMAIL = os.environ['CONTACT_US_EMAIL']
+CONTACT_US_EMAIL = 'alison@kartoza.com'
 
 ELASTIC_MIN_SCORE = 0.5
