--- td_biblio/__init__.py
@@ -1,8 +1,0 @@
-"""TailorDev Biblio
-
-Bibliography management with Django.
-"""
-
-__version__ = '1.2.0'
-
-default_app_config = 'td_biblio.apps.TDBiblioConfig'
