--- td_biblio/apps.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from django.apps import AppConfig
-from django.utils.translation import ugettext_lazy as _
-
-
-class TDBiblioConfig(AppConfig):
-    name = 'td_biblio'
-    verbose_name = _("Bibliography Management")
