--- td_biblio/tests/fixtures/biblio.bib
@@ -1,100 +1,0 @@
-@article{maupetit2007coarse,
-  title={A coarse-grained protein force field for folding and structure prediction},
-  author={Maupetit, Julien and Tuffery, P and Derreumaux, Philippe},
-  journal={Proteins: Structure, Function, and Bioinformatics},
-  volume={69},
-  number={2},
-  pages={394--408},
-  year={2007},
-  publisher={Wiley Online Library}
-}
-
-@article{neron2009mobyle,
-  title={Mobyle: a new full web bioinformatics framework},
-  author={N{\'e}ron, Bertrand and M{\'e}nager, Herv{\'e} and Maufrais, Corinne and Joly, Nicolas and Maupetit, Julien and Letort, S{\'e}bastien and Carrere, S{\'e}bastien and Tuffery, Pierre and Letondal, Catherine},
-  journal={Bioinformatics},
-  volume={25},
-  number={22},
-  pages={3005--3011},
-  year={2009},
-  publisher={Oxford Univ Press}
-}
-
-@article{maupetit2006sabbac,
-  title={SABBAC: online Structural Alphabet-based protein BackBone reconstruction from Alpha-Carbon trace},
-  author={Maupetit, Julien and Gautier, R and Tuff{\'e}ry, Pierre},
-  journal={Nucleic acids research},
-  volume={34},
-  number={suppl 2},
-  pages={W147--W151},
-  year={2006},
-  publisher={Oxford Univ Press}
-}
-
-@article{maupetit2009pep,
-  title={PEP-FOLD: an online resource for de novo peptide structure prediction},
-  author={Maupetit, Julien and Derreumaux, Philippe and Tuffery, Pierre},
-  journal={Nucleic acids research},
-  volume={37},
-  number={suppl 2},
-  pages={W498--W503},
-  year={2009},
-  publisher={Oxford University Press}
-}
-
-@article{maupetit2010fast,
-  title={A fast method for large-scale De Novo peptide and miniprotein structure prediction},
-  author={Maupetit, Julien and Derreumaux, Philippe and Tuff{\'e}ry, Pierre},
-  journal={Journal of computational chemistry},
-  volume={31},
-  number={4},
-  pages={726--738},
-  year={2010},
-  publisher={Wiley Online Library}
-}
-
-@article{schmidtke2010fpocket,
-  title={fpocket: online tools for protein ensemble pocket detection and tracking},
-  author={Schmidtke, Peter and Le Guilloux, Vincent and Maupetit, Julien and Tuff{\'e}ry, Pierre},
-  journal={Nucleic acids research},
-  volume={38},
-  number={suppl 2},
-  pages={W582--W589},
-  year={2010},
-  publisher={Oxford University Press}
-}
-
-@article{lagorce2011faf,
-  title={The FAF-Drugs2 server: a multistep engine to prepare electronic chemical compound collections},
-  author={Lagorce, David and Maupetit, Julien and Baell, Jonathan and Sperandio, Olivier and Tuff{\'e}ry, Pierre and Miteva, Maria A and Galons, Herv{\'e} and Villoutreix, Bruno O},
-  journal={Bioinformatics},
-  volume={27},
-  number={14},
-  pages={2018--2020},
-  year={2011},
-  publisher={Oxford University Press}
-}
-
-@article{thevenet2012pep,
-  title={PEP-FOLD: an updated de novo structure prediction server for both linear and disulfide bonded cyclic peptides},
-  author={Th{\'e}venet, Pierre and Shen, Yimin and Maupetit, Julien and Guyon, Fr{\'e}d{\'e}ric and Derreumaux, Philippe and Tuff{\'e}ry, Pierre},
-  journal={Nucleic acids research},
-  volume={40},
-  number={W1},
-  pages={W288--W293},
-  year={2012},
-  publisher={Oxford University Press}
-}
-
-@article{tudor2012isotype,
-  title={Isotype modulates epitope specificity, affinity, and antiviral activities of anti--HIV-1 human broadly neutralizing 2F5 antibody},
-  author={Tudor, Daniela and Yu, Huifeng and Maupetit, Julien and Drillet, Anne-Sophie and Bouceba, Tahar and Schwartz-Cornil, Isabelle and Lopalco, Lucia and Tuffery, Pierre and Bomsel, Morgane},
-  journal={Proceedings of the National Academy of Sciences},
-  volume={109},
-  number={31},
-  pages={12680--12685},
-  day={31},
-  month={jul},
-  year={2012},
-  publisher={National Acad Sciences}
-}
