--- td_biblio/tests/fixtures/entries.py
@@ -1,59 +1,0 @@
-# Entries from public databases
-
-# PubMed identifiers
-PMIDs = [
-    '16844979',
-    '17600832',
-    '18182455',
-    '19433514',
-    '19569182',
-    '19689959',
-    '20478829',
-    '20972248',
-    '21636592',
-    '21665924',
-    '22581768',
-    '22723360',
-    '23548772',
-    '23649963',
-    '25005735',
-    '25241742',
-    '25603460',
-    '26400168',
-    '26408185',
-    '26588162',
-    '27095822',
-    '24357323',
-    '25190796',
-]
-
-# Digital Object Identfiers
-DOIs = [
-    '10.1002/humu.22352',
-    '10.1002/jcc.21365',
-    '10.1002/prot.21505',
-    '10.1007/978-3-319-19249-9_36',
-    '10.1016/j.stemcr.2014.07.002',
-    '10.1021/ct500592m',
-    '10.1038/ejhg.2014.127',
-    '10.1073/pnas.1200024109',
-    '10.1093/bioinformatics/btp493',
-    '10.1093/bioinformatics/btr333',
-    '10.1093/cercor/bhw102',
-    '10.1093/nar/gkl289',
-    '10.1093/nar/gkp323',
-    '10.1093/nar/gkq383',
-    '10.1093/nar/gkr410',
-    '10.1093/nar/gks419',
-    '10.1093/nar/gkv960',
-    '10.1109/MEMCOD.2015.7340480',
-    '10.1136/jmg.2010.081356',
-    '10.1145/2675743.2771876',
-    '10.1145/2676585.2676615',
-    '10.1159/000348516',
-    '10.1186/s13059-015-0769-z',
-    '10.1210/jc.2007-0927',
-    '10.1210/jc.2014-4047',
-    '10.1126/science.1241089',
-    '10.1126/science.1255274',
-]
