--- td_biblio/urls.py
@@ -1,19 +1,0 @@
-# -*- coding: utf-8 -*-
-from django.conf.urls import url
-
-from . import views
-
-app_name = 'td_biblio'
-urlpatterns = [
-    # Entry List
-    url(
-        '^$',
-        views.EntryListView.as_view(),
-        name='entry_list'
-    ),
-    url(
-        '^import/$',
-        views.EntryBatchImportView.as_view(),
-        name='import'
-    ),
-]
